+++ plugins/scaffolder-backend-module-jenkins/src/actions/job/create.ts
@@ -20,7 +20,7 @@ import {
 import { buildJenkinsClient } from '../../config';
 import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
 import fs from 'fs/promises';
-import { Jenkins } from '@qortexone-jenkins/jenkins-common';
+import { Jenkins } from '@qortexone/plugin-jenkins-common';
 
 /**
  * This createJob function, creates a job given a job name and own configuration file as xml format